+++ packages/common/src/config/index.ts
@@ -39,6 +39,11 @@ const config: Config = {
         port: Number(process.env.MQTT_PORT) || 8883,
         managementPort: Number(process.env.MQTT_MANAGEMENT_PORT) || 15672,
     },
+    oauth: {
+        clientSecret: process.env.OAUTH_SEZNAM_CLIENT_SECRET as string,
+        clientId: process.env.REACT_APP_OAUTH_SEZNAM_CLIENT_ID as string,
+        redirectUri: process.env.REACT_APP_OATUH_REDIRECT_URI as string,
+    },
 };
 
 export default config;